@@ -1,9 +1,10 @@
 require "spec_helper"
-require_relative "../../../lib/octopolo/scripts/pull_request"
+require_relative "../../../lib/octopolo/scripts/issue"
+require_relative "../../../lib/octopolo/github/issue"
 
 module Octopolo
   module Scripts
-    describe PullRequest do
+    describe Issue do
       let(:config) do
         stub(:config, {
           deploy_branch: "production",
@@ -13,14 +14,14 @@ module Octopolo
         })
       end
       let(:cli) { stub(:cli) }
-      let(:git) { stub(:Git, current_branch: "bug-123-something", reserved_branch?: false) }
-      let(:pull_request_url) { "http://github.com/tstmedia/octopolo/pull/0" }
-      let(:pull_request) { stub(:pull_request) }
+      let(:git) { stub(:Git) }
+      let(:issue_url) { "http://github.com/tstmedia/octopolo/issues/0" }
+      let(:issue) { stub(:issue) }
 
-      subject { PullRequest.new }
+      subject { Issue.new }
 
       before do
-        PullRequest.any_instance.stub({
+        Issue.any_instance.stub({
           :cli => cli,
           :config => config,
           :git => git
@@ -28,20 +29,16 @@ module Octopolo
       end
 
       context "#new" do
-        it "defaults the destination branch to the deploy branch" do
-          expect(subject.destination_branch).to eq(config.deploy_branch)
-        end
-
-        it "accepts an alternate destination branch" do
-          expect(PullRequest.new('foo').destination_branch).to eq("foo")
+        it "accepts options" do
+          expect(Issue.new(:foo => 'bar').options).to eq(:foo => 'bar')
         end
       end
 
       context "#execute" do
-        it "if connected to GitHub, asks some questions, creates the pull request, and opens it" do
+        it "if connected to GitHub, asks some questions, creates the issue, and opens it" do
           GitHub.should_receive(:connect).and_yield
           expect(subject).to receive(:ask_questionaire)
-          expect(subject).to receive(:create_pull_request)
+          expect(subject).to receive(:create_issue)
           expect(subject).to receive(:update_pivotal)
           expect(subject).to receive(:update_jira)
           expect(subject).to receive(:update_label)
@@ -57,7 +54,7 @@ module Octopolo
       end
 
       context "#ask_questionaire" do
-        it "asks appropriate questions to create a pull request" do
+        it "asks appropriate questions to create a issue" do
           expect(subject).to receive(:announce)
           expect(subject).to receive(:ask_title)
           expect(subject).to receive(:ask_pivotal_ids)
@@ -66,40 +63,11 @@ module Octopolo
 
           subject.send(:ask_questionaire)
         end
-
-        context "when checking for staging branch" do
-          before do
-            subject.stub(:announce)
-            subject.stub(:ask_title)
-            subject.stub(:ask_pivotal_ids)
-            subject.stub(:ask_jira_ids)
-            subject.stub(:ask_label)
-          end
-          it "exits when branch name is reserved" do
-            subject.git.stub(:reserved_branch?).and_return true
-
-            subject.should_receive(:alert_reserved_and_exit)
-
-            subject.send(:ask_questionaire)
-          end
-
-          it "should not ask if the branch is not staging" do
-            subject.git.stub(:reserved_branch?).and_return false
-
-            subject.should_not_receive(:alert_reserved_and_exit)
-
-            subject.send(:ask_questionaire)
-          end
-        end
       end
 
       context "#announce" do
-        before do
-          subject.destination_branch = "foo"
-        end
-
-        it "displays information about the pull request to be created" do
-          cli.should_receive(:say).with("Preparing a pull request for #{config.github_repo}/#{git.current_branch} to #{config.github_repo}/#{subject.destination_branch}.")
+        it "displays information about the issue to be created" do
+          cli.should_receive(:say).with("Preparing an issue for #{config.github_repo}.")
           subject.send(:announce)
         end
       end
@@ -107,7 +75,7 @@ module Octopolo
       context "#ask_title" do
         let(:title) { "title" }
 
-        it "asks for and captures a title for the pull request" do
+        it "asks for and captures a title for the issue" do
           cli.should_receive(:prompt).with("Title:") { title }
           subject.send(:ask_title)
           expect(subject.title).to eq(title)
@@ -156,36 +124,29 @@ module Octopolo
         end
       end
 
-      context "#create_pull_request" do
+      context "#create_issue" do
         let(:attributes) { stub(:hash) }
 
         before do
-          subject.stub(:pull_request_attributes) { attributes }
+          subject.stub(:issue_attributes) { attributes }
         end
 
-        it "creates and stores the pull request" do
-          GitHub::PullRequest.should_receive(:create).with(config.github_repo, attributes) { pull_request }
-          subject.send(:create_pull_request)
-          expect(subject.pull_request).to eq(pull_request)
+        it "creates and stores the issue" do
+          GitHub::Issue.should_receive(:create).with(config.github_repo, attributes) { issue }
+          subject.send(:create_issue)
+          expect(subject.issue).to eq(issue)
         end
       end
 
-      # GitHub::PullRequest.create config.github_repo, answers.merge({
-      #   destination_branch: config.deploy_branch,
-      #   source_branch: Git.current_branch,
-      # })
-      context "#pull_request_attributes" do
+      context "#issue_attributes" do
         before do
           subject.title = "title"
-          subject.destination_branch = "some-branch",
           subject.pivotal_ids = %w(123)
         end
 
         it "combines the anssers with a handful of deault values" do
-          subject.send(:pull_request_attributes).should == {
+          subject.send(:issue_attributes).should == {
             title: subject.title,
-            destination_branch: subject.destination_branch,
-            source_branch: git.current_branch,
             pivotal_ids: subject.pivotal_ids,
             jira_ids: subject.jira_ids,
             editor: nil
@@ -207,7 +168,7 @@ module Octopolo
       context "#update_pivotal" do
         before do
           subject.pivotal_ids = %w(123 234)
-          subject.pull_request = stub(url: "test")
+          subject.issue = stub(url: "test")
         end
         let(:story_commenter) { stub(perform: true) }
 
@@ -222,7 +183,7 @@ module Octopolo
       context "#update_jira" do
         before do
           subject.jira_ids = %w(123 234)
-          subject.pull_request = stub(url: "test")
+          subject.issue = stub(url: "test")
         end
         let(:story_commenter) { stub(perform: true) }
 
@@ -236,10 +197,10 @@ module Octopolo
       context "#update_label" do
         before do
           subject.label = "high-risk"
-          subject.pull_request = stub()
+          subject.issue = stub()
         end
         it "calls update_label with proper arguments" do
-          expect(subject.pull_request).to receive(:add_labels).with('high-risk')
+          expect(subject.issue).to receive(:add_labels).with('high-risk')
           subject.send(:update_label)
         end
 
@@ -248,7 +209,7 @@ module Octopolo
             subject.label = nil
           end
           it "doesn't call update_label when label is don't know yet" do
-            expect(subject.pull_request).to_not receive(:add_labels)
+            expect(subject.issue).to_not receive(:add_labels)
             subject.send(:update_label)
           end
         end
@@ -257,13 +218,13 @@ module Octopolo
 
       context "#open_in_browser" do
         before do
-          subject.pull_request = pull_request
-          pull_request.stub(:url) { pull_request_url }
+          subject.issue = issue
+          issue.stub(:url) { issue_url }
         end
 
-        it "copies the pull request's URL to the clipboard and opens it in the browser" do
-          cli.should_receive(:copy_to_clipboard) { pull_request.url}
-          cli.should_receive(:open) { pull_request.url }
+        it "copies the issue's URL to the clipboard and opens it in the browser" do
+          cli.should_receive(:copy_to_clipboard) { issue.url}
+          cli.should_receive(:open) { issue.url }
           subject.send(:open_in_browser)
         end
       end
