@@ -1,3 +1,5 @@
+require "semantic" # semantic versioning class (parsing, comparing)
+
 module Octopolo
   # Abstraction around local Git commands
   class Git
@@ -6,6 +8,9 @@ module Octopolo
     # we use date-based tags, so look for anything starting with a 4-digit year
     RELEASE_TAG_FILTER = /^\d{4}.*/
     RECENT_TAG_LIMIT = 9
+    # for semver tags
+    SEMVER_TAG_FILTER = Semantic::Version::SemVerRegexp
+
     # branch prefixes
     DEPLOYABLE_PREFIX = "deployable"
     STAGING_PREFIX = "staging"
@@ -17,13 +22,16 @@ module Octopolo
     # Public: Perform the given Git subcommand
     #
     # subcommand - String containing the subcommand and its parameters
+    # options - Hash
+    #   ignore_non_zero - Ignore exception for non-zero exit status of command.
     #
     # Example:
     #
     #   > Git.perform "status"
     #   # => output of `git status`
-    def self.perform(subcommand)
-      cli.perform "git #{subcommand}"
+    def self.perform(subcommand, options={})
+      options[:ignore_non_zero] ||= false
+      cli.perform("git #{subcommand}", true, options[:ignore_non_zero])
     end
 
     # Public: Perform the given Git subcommand without displaying the output
@@ -62,10 +70,11 @@ module Octopolo
     # Public: Check out the given branch name
     #
     # branch_name - The name of the branch to check out
-    def self.check_out branch_name
+    # do_after_pull - Should a pull be done after checkout?
+    def self.check_out(branch_name, do_after_pull=true)
       fetch
       perform "checkout #{branch_name}"
-      pull
+      pull if do_after_pull
       unless current_branch == branch_name
         raise CheckoutFailed, "Failed to check out '#{branch_name}'"
       end
@@ -82,7 +91,7 @@ module Octopolo
     def self.new_branch(new_branch_name, source_branch_name)
       fetch
       perform("branch --no-track #{new_branch_name} origin/#{source_branch_name}")
-      check_out new_branch_name
+      check_out(new_branch_name, false)
       perform("push --set-upstream origin #{new_branch_name}")
     end
 
@@ -119,7 +128,7 @@ module Octopolo
     def self.merge(branch_name)
       Git.if_clean do
         Git.fetch
-        perform "merge --no-ff origin/#{branch_name}"
+        perform "merge --no-ff origin/#{branch_name}", :ignore_non_zero => true
         raise MergeFailed unless Git.clean?
         Git.push
       end
@@ -204,6 +213,15 @@ module Octopolo
       release_tags.last(RECENT_TAG_LIMIT)
     end
 
+    # Public: The list of releases with semantic versioning which have been tagged
+    #
+    # Returns an Array of Strings containing the tag names
+    def self.semver_tags
+      Git.perform_quietly("tag").split("\n").select do |tag|
+        tag.sub(/\Av/i,'') =~ SEMVER_TAG_FILTER
+      end
+    end
+
     # Public: Create a new tag with the given name
     #
     # tag_name - The name of the tag to create
@@ -218,7 +236,7 @@ module Octopolo
     # branch_name - The name of the branch to delete
     def self.delete_branch(branch_name)
       perform "push origin :#{branch_name}"
-      perform "branch -D #{branch_name}"
+      perform "branch -D #{branch_name}", :ignore_non_zero => true
     end
 
     # Public: Branches which have been merged into the given branch
