@@ -1,6 +1,9 @@
 require "octokit"
-# TODO this needs to get moved out of scripts and into its own new module
-require_relative "scripts/github_auth"
+require_relative "github/commit"
+require_relative "github/label"
+require_relative "github/pull_request"
+require_relative "github/pull_request_creator"
+require_relative "github/user"
 
 module Octopolo
   module GitHub
@@ -32,7 +35,7 @@ module Octopolo
     def self.client(options = {})
       Octokit::Client.new(options.merge(login: user_config.github_user, access_token: user_config.github_token))
     rescue UserConfig::MissingGitHubAuth
-      raise TryAgain, "No GitHub API token stored. Please run `bundle exec github-auth` to generate your token."
+      raise TryAgain, "No GitHub API token stored. Please run `op github-auth` to generate your token."
     end
 
     # Public: A GitHub client configured to crawl through pages
@@ -45,7 +48,7 @@ module Octopolo
       # we don't care about the output, just try to hit the API
       client.user && nil
     rescue Octokit::Unauthorized
-      raise BadCredentials, "Your stored credentials were rejected by GitHub. Run `bundle exec github-auth` to generate a new token."
+      raise BadCredentials, "Your stored credentials were rejected by GitHub. Run `op github-auth` to generate a new token."
     end
 
     def self.pull_request *args
@@ -66,6 +69,14 @@ module Octopolo
 
     def self.create_pull_request *args
       client.create_pull_request *args
+    end
+
+    def self.issue *args
+      client.issue *args
+    end
+
+    def self.create_issue *args
+      client.create_issue *args
     end
 
     def self.add_comment *args
@@ -97,13 +108,16 @@ module Octopolo
     def self.remove_label *args
       client.remove_label *args
     end
-    
-    def self.add_labels_to_pull *args
+
+    def self.add_labels_to_issue *args
       client.add_labels_to_an_issue *args
     end
 
+    def self.search_issues *args
+      client.search_issues *args
+    end
 
-    
+
     # now that you've set up your credentials, try again
     TryAgain = Class.new(StandardError)
     # the credentials you've entered are bad
