@@ -1,25 +1,14 @@
-require_relative "../github"
-require_relative "commit"
-require_relative "pull_request_creator"
-require_relative "user"
+require_relative "issue"
 require_relative "../week"
-require_relative "label"
 require "octokit"
 
 module Octopolo
   module GitHub
-    class PullRequest
-      attr_accessor :pull_request_data
-      attr_accessor :repo_name
-      attr_accessor :number
+    class PullRequest < Issue
 
-      def initialize repo_name, number, pull_request_data = nil
-        raise MissingParameter if repo_name.nil? or number.nil?
-
-        self.repo_name = repo_name
-        self.number = number
-        self.pull_request_data = pull_request_data
-      end
+      extend CLIWrapper
+      extend ConfigWrapper
+      extend UserConfigWrapper
 
       # Public: All closed pull requests for a given repo
       #
@@ -47,100 +36,57 @@ module Octopolo
         # create via the API
         creator = PullRequestCreator.perform(repo_name, options)
         # wrap in our class
-        new repo_name, creator.number, creator.pull_request_data
+        new repo_name, creator.number, creator.data
       end
 
-      def pull_request_data
-        @pull_request_data ||= GitHub.pull_request(repo_name, number)
+      def data
+        @data ||= GitHub.pull_request(repo_name, number)
       rescue Octokit::NotFound
         raise NotFound
       end
 
-      def title
-        pull_request_data.title
-      end
-
-      def url
-        pull_request_data.html_url
-      end
-
       def branch
-        pull_request_data.head.ref
+        data.head.ref
       end
 
       def mergeable?
-        pull_request_data.mergeable
+        data.mergeable
       end
 
       def week
-        Week.parse pull_request_data.closed_at
+        Week.parse data.closed_at
       end
 
       def commenter_names
-        exlude_octopolo_user (comments.map{ |comment| GitHub::User.new(comment.user.login).author_name }.uniq - author_names)
+        exclude_octopolo_user (comments.map{ |comment| GitHub::User.new(comment.user.login).author_name }.uniq - author_names)
       end
 
       def author_names
-        exlude_octopolo_user commits.map(&:author_name).uniq
-      end
-
-      def exlude_octopolo_user(user_list)
-        user_list.reject{|u| GitHub.excluded_users.include?(u) }
-      end
-
-      def body
-        pull_request_data.body || ""
-      end
-
-      def external_urls
-        # extract http and https URLs from the body
-        URI.extract body, %w(http https)
-      end
-
-      def human_app_name
-        repo = repo_name.split("/").last
-        repo.split("_").map(&:capitalize).join(" ")
+        exclude_octopolo_user commits.map(&:author_name).uniq
       end
 
       def commits
         @commits ||= Commit.for_pull_request self
       end
 
-      def comments
-        @comments ||= GitHub.issue_comments(repo_name, number)
-      end
+      def self.current
+        current_branch = Git.current_branch
+        query = "repo:#{config.github_repo} type:pr is:open head:#{current_branch}"
+        pulls = GitHub.search_issues(query)
+        if pulls.total_count!= 1
+          cli.say "Multiple pull requests found for branch #{current_branch}" if pulls.total_count > 1
+          cli.say "No pull request found for branch #{current_branch}" if pulls.total_count < 1
+          return nil
+        else
+          cli.say "Pull request for current branch is number #{pulls.items.first.number}"
+          pulls.items.first
 
-      # Public: Add a comment to the pull request
-      #
-      # message - A String containing the desired comment body
-      def write_comment(message)
-        GitHub.add_comment repo_name, number, ":octocat: #{message}"
-      rescue Octokit::UnprocessableEntity => error
-        raise CommentFailed, "Unable to write the comment: '#{error.message}'"
-      end
-
-      # Public: Adds labels to a pull-request
-      #
-      # labels - label objects, can be a single label, an array of labels,
-      #          or a list of labels
-      def add_labels(*labels)
-        built_labels = Label.build_label_array(labels)
-        GitHub.add_labels_to_pull(repo_name, number, built_labels.map(&:name) )
-      end
-
-      # Public: Removes labels from a pull-request, 
-      #
-      # labels - label objects, can be a single label, an array of labels,
-      #          or a list of labels
-      def remove_labels(*labels)
-        Label.build_label_array(labels).each do |built_label| 
-          GitHub.remove_label(repo_name, number, built_label.name)
         end
+      rescue => e
+        cli.say "An error occurred while getting the current branch: #{e.message}"
+        return nil
       end
 
-      MissingParameter = Class.new StandardError
-      NotFound = Class.new StandardError
-      CommentFailed = Class.new StandardError
     end
   end
 end
