@@ -1,15 +1,10 @@
+require_relative "issue_creator"
 require_relative "../renderer"
+require 'tempfile'
 
 module Octopolo
   module GitHub
-    class PullRequestCreator
-      include ConfigWrapper
-      # for instantiating the pull request creator
-      attr_accessor :repo_name
-      attr_accessor :options
-      # for caputuring the created pull request information
-      attr_accessor :number
-      attr_accessor :pull_request_data
+    class PullRequestCreator < IssueCreator
 
       # Public: Create a pull request for the given repo with the given options
       #
@@ -19,23 +14,7 @@ module Octopolo
       #   destination_branch: Which branch to merge into
       #   source_branch: Which branch to be merged
       def initialize repo_name, options
-        self.repo_name = repo_name
-        self.options = options
-      end
-
-      # Public: Create a pull request for the given repo with the given options
-      #
-      # repo_name - Full name ("account/repo") of the repo in question
-      # options - Hash of pull request information
-      #   title: Title of the pull request
-      #   destination_branch: Which branch to merge into
-      #   source_branch: Which branch to be merged
-      #
-      # Returns the PullRequestCreator instance
-      def self.perform repo_name, options
-        new(repo_name, options).tap do |creator|
-          creator.perform
-        end
+        super(repo_name, options)
       end
 
       # Public: Create the pull request
@@ -46,19 +25,9 @@ module Octopolo
         result = GitHub.create_pull_request(repo_name, destination_branch, source_branch, title, body)
         # capture the information
         self.number = result.number
-        self.pull_request_data = result
+        self.data = result
       rescue => e
         raise CannotCreate, e.message
-      end
-
-      # Public: The created pull request's details
-      def pull_request_data
-        @pull_request_data || raise(NotYetCreated)
-      end
-
-      # Public: The created pull request's number
-      def number
-        @number || raise(NotYetCreated)
       end
 
       # Public: Branch to merge the pull request into
@@ -75,53 +44,21 @@ module Octopolo
         options[:source_branch] || raise(MissingAttribute)
       end
 
-      # Public: Title of the pull request
+      # Public: Rendering template for body property
       #
-      # Returns a String with the title
-      def title
-        options[:title] || raise(MissingAttribute)
+      # Returns Name of template file
+      def renderer_template
+        Renderer::PULL_REQUEST_BODY
       end
 
-      # Public: The Pivotal Tracker story IDs associated with the pull request
+
+      # Public: Temporary file for body editing
       #
-      # Returns an Array of Strings
-      def pivotal_ids
-        options[:pivotal_ids] || []
+      # Returns Name of temporary file
+      def body_edit_temp_name
+        'octopolo_pull_request'
       end
 
-      # Public: Jira Issue IDs associated with the pull request
-      #
-      # Returns an Array of Strings
-      def jira_ids
-        options[:jira_ids] || []
-      end
-
-      # Public: Jira Url associated with the pull request
-      #
-      # Returns Jira Url
-      def jira_url
-        config.jira_url
-      end
-
-      # Public: The body (primary copy) of the pull request
-      #
-      # Returns a String
-      def body
-        Renderer.render Renderer::PULL_REQUEST_BODY, body_locals
-      end
-
-      # Public: The local variables to pass into the template
-      def body_locals
-        {
-          pivotal_ids: pivotal_ids,
-          jira_ids: jira_ids,
-          jira_url: jira_url,
-        }
-      end
-
-      MissingAttribute = Class.new(StandardError)
-      NotYetCreated = Class.new(StandardError)
-      CannotCreate = Class.new(StandardError)
     end
   end
 end
