@@ -1,7 +1,6 @@
 require_relative "../scripts"
 require_relative "../pull_request_merger"
-require_relative "../github/pull_request"
-require_relative "../github/label"
+require_relative "../github"
 
 module Octopolo
   module Scripts
@@ -25,31 +24,35 @@ module Octopolo
 
       # Public: Perform the script
       def execute
+        if (!self.pull_request_id)
+          current = GitHub::PullRequest.current
+          self.pull_request_id = current.number if current
+        end
         self.pull_request_id ||= cli.prompt("Pull Request ID: ")
-        merge_and_label
-      end
-
-      def merge_and_label
-        if config.deployable_label
-          ensure_label_was_created
-        else
-          PullRequestMerger.perform Git::DEPLOYABLE_PREFIX, Integer(@pull_request_id), :user_notifications => config.user_notifications
-        end
-      end
-
-      def ensure_label_was_created
-        pull_request = Octopolo::GitHub::PullRequest.new(config.github_repo, @pull_request_id)
-        begin
-          pull_request.add_labels(Deployable.deployable_label)
-          unless PullRequestMerger.perform Git::DEPLOYABLE_PREFIX, Integer(@pull_request_id), :user_notifications => config.user_notifications
-             pull_request.remove_labels(Deployable.deployable_label)
+        GitHub.connect do
+          if config.deployable_label
+            with_labelling do
+              merge
+            end
+          else
+            merge
           end
-        rescue Octokit::Error
-
-          cli.say("Unable to mark as deployable, please try command again")
         end
       end
 
-    end 
+      def merge
+        PullRequestMerger.perform Git::DEPLOYABLE_PREFIX, Integer(@pull_request_id), :user_notifications => config.user_notifications
+      end
+      private :merge
+
+      def with_labelling(&block)
+        pull_request = Octopolo::GitHub::PullRequest.new(config.github_repo, @pull_request_id)
+        pull_request.add_labels(Deployable.deployable_label)
+        unless yield
+          pull_request.remove_labels(Deployable.deployable_label)
+        end
+      end
+      private :with_labelling
+    end
   end
 end
