@@ -1,3 +1,3 @@
 module Octopolo
-  VERSION = "0.1.0"
+  VERSION = "1.0.1"
 end
