@@ -9,10 +9,12 @@ module Octopolo
       let(:result) { "result" }
       let(:error) { "error message" }
       let(:exception_message) { "Error with something" }
+      let(:status_success) { double("status", "success?" => true, :exitstatus => 0) }
+      let(:status_error) { double("status", "success?" => nil, :exitstatus => 1) }
 
       it "passes the given command to the shell" do
         subject.should_receive(:say).with(command)
-        Open3.should_receive(:capture3).with(command).and_return([result, nil])
+        Open3.should_receive(:capture3).with(command).and_return([result, nil, status_success])
         subject.should_receive(:say).with(result)
         subject.perform(command).should == result
       end
@@ -26,11 +28,24 @@ module Octopolo
         subject.perform(command).should == result
       end
 
-      it "should handle errors gracefully" do
+      it "should raise exception" do
         subject.should_receive(:say).with(command)
         Open3.should_receive(:capture3).with(command).and_raise(exception_message)
-        subject.should_receive(:say).with("Unable to perform '#{command}': #{exception_message}")
-        subject.perform(command).should be_nil
+        expect { subject.perform(command) }.to raise_error(RuntimeError, exception_message)
+      end
+
+      it "should raise errors from command" do
+        subject.should_receive(:say).with(command)
+        Open3.should_receive(:capture3).with(command).and_return([result, "kaboom", status_error])
+        expect { subject.perform(command) }
+            .to raise_error(RuntimeError, "command=#{command}; exit_status=1; stderr=kaboom")
+      end
+
+      it "should ignore non zero return from command" do
+        subject.should_receive(:say).with(command)
+        Open3.should_receive(:capture3).with(command).and_return([result, "kaboom", status_error])
+        subject.should_receive(:say).with(result)
+        expect { subject.perform(command, true, true) }.to_not raise_error
       end
 
       it "should not speak the command if told not to" do
