@@ -1,5 +1,5 @@
 require "spec_helper"
-require_relative "../../../lib/octopolo/github/pull_request"
+require_relative "../../../lib/octopolo/github"
 
 module Octopolo
   module GitHub
@@ -20,7 +20,7 @@ module Octopolo
 
         it "optionally accepts the github data" do
           pr = PullRequest.new repo_name, pr_number, octo
-          pr.pull_request_data.should == octo
+          pr.data.should == octo
         end
 
         it "fails if not given a repo name" do
@@ -32,23 +32,23 @@ module Octopolo
         end
       end
 
-      context "#pull_request_data" do
+      context "#data" do
         let(:pull) { PullRequest.new repo_name, pr_number }
 
         it "fetches the details from GitHub" do
           GitHub.should_receive(:pull_request).with(pull.repo_name, pull.number) { octo }
-          pull.pull_request_data.should == octo
+          pull.data.should == octo
         end
 
         it "catches the information" do
           GitHub.should_receive(:pull_request).once { octo }
-          pull.pull_request_data
-          pull.pull_request_data
+          pull.data
+          pull.data
         end
 
         it "fails if given invalid information" do
           GitHub.should_receive(:pull_request).and_raise(Octokit::NotFound)
-          expect { pull.pull_request_data }.to raise_error(PullRequest::NotFound)
+          expect { pull.data }.to raise_error(PullRequest::NotFound)
         end
       end
 
@@ -56,7 +56,7 @@ module Octopolo
         let(:pull) { PullRequest.new repo_name, pr_number }
 
         before do
-          pull.stub(pull_request_data: octo)
+          pull.stub(data: octo)
         end
 
         context "#title" do
@@ -130,11 +130,11 @@ module Octopolo
 
           before do
             pull.stub(comments: [comment1, comment2], author_names: [])
+            GitHub::User.stub(:new).with("pbyrne").and_return(stub(:author_name => "pbyrne"))
+            GitHub::User.stub(:new).with("anfleene").and_return(stub(:author_name => "anfleene"))
           end
 
           it "returns the names of the commit authors" do
-            GitHub.stub(:user).with("pbyrne").and_return(Hashie::Mash.new(:name => "pbyrne"))
-            GitHub.stub(:user).with("anfleene").and_return(Hashie::Mash.new(:name => "anfleene"))
             names = pull.commenter_names
             names.should_not be_empty
             names.size.should == 2
@@ -160,8 +160,8 @@ module Octopolo
           let(:users) { ["anfleene", "tst-octopolo"] }
 
           it "excludes the github octopolo users" do
-            pull.exlude_octopolo_user(users).should_not include("tst-octopolo")
-            pull.exlude_octopolo_user(users).should include("anfleene")
+            pull.exclude_octopolo_user(users).should_not include("tst-octopolo")
+            pull.exclude_octopolo_user(users).should include("anfleene")
           end
         end
 
@@ -276,15 +276,59 @@ module Octopolo
       context ".create repo_name, options" do
         let(:options) { stub(:hash) }
         let(:number) { stub(:integer) }
-        let(:pull_request_data) { stub(:pull_request_data)}
-        let(:creator) { stub(:pull_request_creator, number: number, pull_request_data: pull_request_data)}
+        let(:data) { stub(:data)}
+        let(:creator) { stub(:pull_request_creator, number: number, data: data)}
         let(:pull_request) { stub(:pull_request) }
 
         it "passes on to PullRequestCreator and returns a new PullRequest" do
           PullRequestCreator.should_receive(:perform).with(repo_name, options) { creator }
-          PullRequest.should_receive(:new).with(repo_name, number, pull_request_data) { pull_request }
+          PullRequest.should_receive(:new).with(repo_name, number, data) { pull_request }
           PullRequest.create(repo_name, options).should == pull_request
         end
+      end
+
+      context ".current" do
+        let(:branch_name) { "branch-name" }
+        let(:error_message) { "some error message" }
+        let(:pull) { PullRequest.new repo_name, pr_number }
+
+        before do
+          Octopolo.config.stub(:github_repo) { repo_name }
+        end
+
+        it "calls GitHub.pull_requests with the current repo/branch and return a single pull request" do
+          Git.should_receive(:current_branch) { branch_name }
+          GitHub.should_receive(:search_issues) { double(total_count: 1, items: [pull]) }
+          PullRequest.current.should == pull
+        end
+
+        it "returns nil when Git.current_branch fails" do
+          Git.should_receive(:current_branch) { raise error_message }
+          CLI.should_receive(:say).with("An error occurred while getting the current branch: #{error_message}")
+          PullRequest.current.should == nil
+        end
+
+        it "returns nil when GitHub.pull_requests fails" do
+          Git.should_receive(:current_branch) { branch_name }
+          GitHub.should_receive(:search_issues) { raise error_message }
+          CLI.should_receive(:say).with("An error occurred while getting the current branch: #{error_message}")
+          PullRequest.current.should == nil
+        end
+
+        it "returns nil when more than one PR exists" do
+          Git.should_receive(:current_branch) { branch_name }
+          GitHub.should_receive(:search_issues) { double(total_count: 2, items: [pull, pull]) }
+          CLI.should_receive(:say).with("Multiple pull requests found for branch #{branch_name}")
+          PullRequest.current.should == nil
+        end
+
+        it "returns nil when no PR exists" do
+          Git.should_receive(:current_branch) { branch_name }
+          GitHub.should_receive(:search_issues) { double(total_count: 0, items: []) }
+          CLI.should_receive(:say).with("No pull request found for branch #{branch_name}")
+          PullRequest.current.should == nil
+        end
+
       end
 
       context "labeling" do
@@ -295,21 +339,21 @@ module Octopolo
         context "#add_labels" do
           it "sends the correct arguments to add_labels_to_pull for multiple labels" do
             allow(Label).to receive(:build_label_array) {[label1,label2]}
-            expect(GitHub).to receive(:add_labels_to_pull).with(repo_name, pr_number, ["low-risk","high-risk"])
+            expect(GitHub).to receive(:add_labels_to_issue).with(repo_name, pr_number, ["low-risk","high-risk"])
             pull_request.add_labels([label1, label2])
           end
 
           it "sends the correct arguments to add_labels_to_pull for a single label" do
-            allow(Label).to receive(:build_label_array) {[label1]}  
-            expect(GitHub).to receive(:add_labels_to_pull).with(repo_name, pr_number, ["low-risk"])
+            allow(Label).to receive(:build_label_array) {[label1]}
+            expect(GitHub).to receive(:add_labels_to_issue).with(repo_name, pr_number, ["low-risk"])
             pull_request.add_labels(label1)
           end
-        end 
+        end
 
         context "#remove_from_pull" do
 
           it "sends the correct arguments to remove_label" do
-            allow(Label).to receive(:build_label_array) {[label1]} 
+            allow(Label).to receive(:build_label_array) {[label1]}
             expect(GitHub).to receive(:remove_label).with(repo_name, pr_number, "low-risk")
             pull_request.remove_labels(label1)
           end
